feat: delete original telegraf.go
