+++ cmd/telegraf/telegraf_windows_test.go
