--- cmd/telegraf_cliv2/telegraf_posix.go
@@ -1,9 +1,0 @@
-//go:build !windows
-// +build !windows
-
-package main
-
-func (a *AgentManager) Run() error {
-	stop = make(chan struct{})
-	return a.reloadLoop()
-}
